🚧 [FE] cont: 캘린더 패치 기능 추가 , supabase.ts 파일 수정
