+++ components/supabase.ts
@@ -1,4 +1,11 @@
+// supabaseClient.ts
 import { createClient } from "@supabase/supabase-js";
-const supabaseUrl = "https://jyrfxniwlrpwpcdjivbe.supabase.co";
-const supabaseKey = process.env.SUPABASE_KEY;
-export const supabase = createClient(supabaseUrl, supabaseKey as string);
+
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL as string;
+const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_KEY as string;
+
+if (!supabaseUrl || !supabaseKey) {
+  throw new Error("Supabase URL and Key 필수여야 해용.");
+}
+
+export const supabase = createClient(supabaseUrl, supabaseKey);
