+++ package.json
@@ -17,7 +17,8 @@
     "react-hook-form": "^7.51.5",
     "react-kakao-maps-sdk": "^1.1.27",
     "recoil": "^0.7.7",
-    "sass": "^1.77.4"
+    "sass": "^1.77.4",
+    "supabase": "^1.172.2"
   },
   "devDependencies": {
     "@types/node": "^20",
